Se agrgaron nuevos archivos paquete 2.5 de prueba y paquete 3

+++ app/Models/NumberSequence.php
@@ -17,17 +17,15 @@ class NumberSequence extends Model
      */
     public static function next(string $type, int $userId): string
     {
-        $year = now()->year;
+        $year = (int) now()->year;
 
         $prefix = match ($type) {
             'budget'  => 'PRES',
-            'invoice' => 'FAC', // coherente con el resto del proyecto
+            'invoice' => 'FAC',
             default   => strtoupper($type),
         };
 
         return DB::transaction(function () use ($type, $userId, $year, $prefix) {
-            // Localiza o crea la fila del contador para (user_id, type, year)
-            // y la bloquea para evitar carreras.
             $row = static::where([
                     'user_id' => $userId,
                     'type'    => $type,
@@ -43,14 +41,11 @@ class NumberSequence extends Model
                     'year'    => $year,
                     'last'    => 0,
                 ]);
-                // Nota: al crear ya estamos dentro de la transacción.
             }
 
-            // Incrementa el contador y guarda
             $row->last = (int) $row->last + 1;
             $row->save();
 
-            // Formatea la secuencia a 4 dígitos
             $seq = str_pad((string) $row->last, 4, '0', STR_PAD_LEFT);
 
             return "{$prefix}-{$year}-{$seq}";
